formatting updates, some typing fixes
